Working on making it so that the original sequence indices in a link_to_coarse are represented with 64-bit integers

+++ src/compressed.h
@@ -14,24 +14,24 @@
 
 struct cbp_link_to_coarse {
     char *diff;
-    int64_t coarse_seq_id;
-    int16_t original_start;
-    int16_t original_end;
-    int16_t coarse_start;
-    int16_t coarse_end;
+    uint16_t coarse_seq_id;
+    uint64_t original_start;
+    uint64_t original_end;
+    uint16_t coarse_start;
+    uint16_t coarse_end;
     struct cbp_link_to_coarse *next;
 };
 
 struct cbp_link_to_coarse *
 cbp_link_to_coarse_init(int32_t coarse_seq_id,
-                        int16_t original_start, int16_t original_end,
-                        int16_t coarse_start, int16_t coarse_end,
+                        uint64_t original_start, uint64_t original_end,
+                        uint16_t coarse_start, uint16_t coarse_end,
                         struct cbp_alignment alignment, bool dir);
 
 struct cbp_link_to_coarse *
 cbp_link_to_coarse_init_nodiff(int32_t coarse_seq_id,
-                               int16_t original_start, int16_t original_end,
-                               int16_t coarse_start, int16_t coarse_end,
+                               uint64_t original_start, uint64_t original_end,
+                               uint16_t coarse_start, uint16_t coarse_end,
                                bool dir);
 
 void
